project update

+++ mundo01/desafios/d008.py
@@ -1,3 +1,3 @@
-m = int(input('Qual distância em metros: '))
+m = float(input('Qual distância em metros: '))
 print('{} metros são {} centímetros'.format(m, m*100))
 print('{} metros são {} milímetros'.format(m, m*1000))
